Item deletion.

+++ trunk/src/com/scalar/freequent/action/ManageGroupsAction.java
@@ -20,6 +20,7 @@ import com.scalar.core.ScalarServiceException;
 import com.scalar.core.ScalarValidationException;
 import com.scalar.core.jdbc.DAOFactory;
 import com.scalar.core.util.MsgObjectUtil;
+import com.scalar.core.util.GUID;
 import com.scalar.core.service.ServiceFactory;
 import com.scalar.core.response.Response;
 
@@ -131,29 +132,34 @@ public class ManageGroupsAction extends AbstractActionController {
 	}
 
     public void delete(Request request, Object command, Map<String, Object> data) throws ScalarActionException {
+        Map<String, String> params = new HashMap<String, String>();
         boolean isDeleteed = false;
+        
 		GroupDataService groupDataService = ServiceFactory.getService(GroupDataService.class, request);
         ItemDataService itemDataService = ServiceFactory.getService(ItemDataService.class, request);
+        
         String id = request.getParameter(Constants.ITEM_ID_ATTRIBUTE);
         String groupName = request.getParameter(Item.PARAM_NAME);
-        
-        Map<String, String> params = new HashMap<String, String>();
         params.put(Item.PARAM_GROUP, groupName);
-        List<Item> items= itemDataService.search(params);
 
         if (logger.isDebugEnabled()) {
 			logger.debug("Deleted group id: " + id);
             logger.debug("Deleted group name: " + groupName);
 		}
-        
+
+        if (StringUtil.isEmpty(id)) {
+			throw ScalarActionException.create(MsgObjectUtil.getMsgObject(ActionResource.BASE_NAME, ActionResource.PARAM_REQUIRED, id), null);
+		}
+
+		if (!GUID.isValid(id)) {
+			throw ScalarActionException.create(MsgObjectUtil.getMsgObject(ActionResource.BASE_NAME, ActionResource.INVALID_PARAM_VALUE, groupName, id), null);
+		}
+
+        List<Item> items= itemDataService.search(params);
         if (items.size() > 0) {
-            throw ScalarServiceException.create(MsgObjectUtil.getMsgObject(ServiceResource.BASE_NAME, ServiceResource.UNABLE_TO_DELETE_ASSOCIATED_GROUP, ObjectType.GROUP, groupName), null);
+            throw ScalarActionException.create(MsgObjectUtil.getMsgObject(ActionResource.BASE_NAME, ActionResource.UNABLE_TO_DELETE_ASSOCIATED_GROUP, ObjectType.GROUP, groupName), null);
         } else {
-            try {
-                isDeleteed = groupDataService.remove(id);
-            } catch (ScalarServiceException e) {
-			    throw getActionException(e);
-		    }
+            isDeleteed = groupDataService.remove(id);
         }
         Map<String, Boolean> succes = new HashMap<String, Boolean>();
         succes.put("success", isDeleteed);
